Cleaned up configure.in to correctly define (or not) flags like HAVE_GETADDRINFO_AI_ADDRCONFIG.
The problem was that AC_DEFINE([blah], 0) will still put a #define into
config.h, albeit with a value of 0, but this will cause #ifdef's to return true
in the code.

Instead, AC_DEFINE should only be used when we actually want to define the
variable.  autoheader will still pick up the define regardless of where it it
in configure.in

Tested: darwin (osx, with ipv6), linux (no ipv6)