Various fixes:
- Fixed some show-stoppers (like calling fatal for the "connected" message ;) )
- Changed verbose messages to be more netcat like
- Cleaned up reverse DNS code, while still ensuring errors in inverse lookup
  is non-fatal.
- Reworked the "double bind" issue:

On systems that do "double binding" for ipv6 listening sockets, binding to an
ipv4 address will fail if the ipv6 address is already bound (and vice versa).
However, ipv6 sockets will still accept connections targeted for the ipv4
equivalent address - but not the other way around.

So, to ensure portability (at least in function), care is taken to always bind
to ipv6 sockets first - that way all connections will still work.  The only
limitation to this approach is that the second bind will be reported as a
failure in verbose mode.  At some point a autoconf test should be written for
this and platforms that do double binding should suppress the error message.