I was thinking over the half close stuff and realised there was a major problem with the implementation.  Whenever a eof was received, the opposing write stream was shutdown - REGARDLESS or any data still in the output buffer.
To rectify this, I've added a method to io_stream called ios_write_eof, which
signals the io_stream that no more data will be added to the output buffer.
When the output buffer is emptied, the io_stream takes care of calling
shutdown for the write stream.

This change stuffs up the very verbose logging a little, but I'm working on a
fix for this.