Cleaned up main. Normalized ca interface - providing additional accessors.
Yes - I've slightly ignored demeter's law for ca wrt the ca_remote_stream and
ca_local_stream accessors (they return 'objects').  But trying to do it any
other way results in all the code that was in main ending up in a function of
the ca and being just as messy as before.  I've taken the view that the ca
is basically just a storage for all our 'attributes', not a representation of
the actual connection, so returning these objects is necessary.