Fixed that hold timeout problem (there was an easy solution actually).
Changed to close the write of the ios that timed out, and the read of the
other one.  This only leaves the write of the other endpoint open - allowing
the data from the timed out endpoint to be written out properly (and any
errors recorded).  Since the write of the timed out endpoint is closed, no
more writes will be attempted and hence no errors can occur.

Eg. In normal mode, when the remote read gets EOF and is closed, then the
remote write will also be closed as well as the local read (stdin).  Only the
local write (stdout) will be left open until the output buffer is empty.

I realised the solution about 2 minutes after going to bed last night - but I
managed to not get up and add it till now.