2nd bugfix (major)
The fix I made to ensure buffer flushing after a hold timeout has a serious
problem.  If the buffer isn't already flushed when the hold timeout is
triggered, it enters a infinite loop(!*$#XB&#@$@#$%#$%).  This is because
finding a hold timeout condition triggers a jump back to the top of the
readwrite loop (after shutting down the appropriate streams) in order to
redetect which streams are ready to be read/written to.  If there is anything
to be read/written then the (max_fd == -1) check will fail, which is followed
by checking the timeouts again....and finding the hold timeout
condition....and jumping to the top....and....and....and....and......

I am annoyed.